Moved all gui classes into respective subpackages.
If you feel there is a better way to divide them, feel free to give a suggestion!
