+++ src/main/java/gui/centerarea/TimelinesGridPane.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.centerarea;
 
 import java.util.ArrayList;
 import javafx.geometry.Bounds;
@@ -6,7 +6,7 @@ import javafx.geometry.Insets;
 
 
 /**
- * Class that represents the grid pane in the scrollable camera timelines.
+ * Class that represents the grid pane in the scrollable camera centerarea.
  */
 public class TimelinesGridPane extends ScrollableGridPane {
