+++ src/main/java/gui/events/DirectorShotCreationEvent.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.events;
 
 import lombok.Getter;
