+++ src/main/java/gui/events/ModalEvent.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.events;
 
 import javafx.event.Event;
 import javafx.event.EventType;
