+++ src/main/java/gui/headerarea/DetailView.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.headerarea;
 
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
