+++ src/main/java/gui/headerarea/NumberTextField.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.headerarea;
 
 
 import javafx.scene.control.TextField;
