+++ src/main/java/gui/headerarea/ToolButton.java
@@ -1,5 +1,6 @@
-package gui;
+package gui.headerarea;
 
+import gui.root.RootHeaderArea;
 import javafx.event.EventHandler;
 import javafx.scene.control.Button;
 import javafx.scene.input.MouseEvent;
