+++ src/main/java/gui/root/RootPane.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.root;
 
 import control.ControllerManager;
 import javafx.application.Application;
