+++ src/test/java/xml/XmlWriterTest.java
@@ -66,7 +66,7 @@ public class XmlWriterTest {
 
         String file = readFile("src/test/java/xml/test_files/test-write.xml", StandardCharsets.UTF_8);
 
-        assertEquals("<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"no\"?><scripting-project description=\"A test project\" seconds-per-count=\"2.0\"><director-timeline description=\"A test director timeline\"><director-shots number-of-shots=\"1\"><director-shot description=\"A test director shot\" instance=\"" + directorShotInstance + "\" name=\"shot-1\"/></director-shots></director-timeline><camera-timelines><camera-timeline description=\"Test camera line 1\"><camera description=\"some camera description\" movement-margin=\"4.0\" name=\"Some camera\"><camera-type description=\"b\" movement-margin=\"4.0\" name=\"a\"/></camera><camera-shots number-of-shots=\"1\"><camera-shot description=\"A camera shot\" instance=\"" + cameraShotInstance1 + "\" name=\"shot-2\"/></camera-shots></camera-timeline><camera-timeline description=\"Test camera line 2\"><camera description=\"another description\" movement-margin=\"5.0\" name=\"Another camera\"><camera-type description=\"d\" movement-margin=\"5.0\" name=\"c\"/></camera><camera-shots number-of-shots=\"1\"><camera-shot description=\"Another camera shot\" instance=\"" + cameraShotInstance2 + "\" name=\"shot-3\"/></camera-shots></camera-timeline></camera-timelines><cameras><camera description=\"Camera description\" movement-margin=\"4.0\" name=\"Camera\"><camera-type description=\"y\" movement-margin=\"4.0\" name=\"x\"/></camera></cameras></scripting-project>", file);
+        assertEquals("<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"no\"?><scripting-project description=\"A test project\" seconds-per-count=\"2.0\"><director-timeline description=\"A test director timeline\"><director-shots number-of-shots=\"1\"><director-shot description=\"A test director shot\" instance=\"" + directorShotInstance + "\" name=\"shot-1\"/></director-shots></director-timeline><camera-centerarea><camera-timeline description=\"Test camera line 1\"><camera description=\"some camera description\" movement-margin=\"4.0\" name=\"Some camera\"><camera-type description=\"b\" movement-margin=\"4.0\" name=\"a\"/></camera><camera-shots number-of-shots=\"1\"><camera-shot description=\"A camera shot\" instance=\"" + cameraShotInstance1 + "\" name=\"shot-2\"/></camera-shots></camera-timeline><camera-timeline description=\"Test camera line 2\"><camera description=\"another description\" movement-margin=\"5.0\" name=\"Another camera\"><camera-type description=\"d\" movement-margin=\"5.0\" name=\"c\"/></camera><camera-shots number-of-shots=\"1\"><camera-shot description=\"Another camera shot\" instance=\"" + cameraShotInstance2 + "\" name=\"shot-3\"/></camera-shots></camera-timeline></camera-centerarea><cameras><camera description=\"Camera description\" movement-margin=\"4.0\" name=\"Camera\"><camera-type description=\"y\" movement-margin=\"4.0\" name=\"x\"/></camera></cameras></scripting-project>", file);
 
     }
